.env 환경 변수 추가

+++ controller/post.mjs
@@ -2,46 +2,57 @@ import * as postRepository from "../data/post.mjs";
 
 // 모든 포스트를 가져오는 함수
 export async function getPosts(req, res, next) {
-  const userid = req.query.userid;
-  const data = await (userid
-    ? postRepository.getAllByUserid(userid)
-    : postRepository.getAll());
-  res.status(200).json(data);
+    const userid = req.query.userid;
+    const data = await (userid
+        ? postRepository.getAllByUserid(userid)
+        : postRepository.getAll());
+    res.status(200).json(data);
 }
 
 // 하나의 포스트를 가져오는 함수
 export async function getPost(req, res, next) {
-  const id = req.params.id;
-  const post = await postRepository.getById(id);
-  if (post) {
-    res.status(200).json(post);
-  } else {
-    res.status(404).json({ message: `${id}의 포스트가 없습니다` });
-  }
+    const id = req.params.id;
+    const post = await postRepository.getById(id);
+    if (post) {
+        res.status(200).json(post);
+    } else {
+        res.status(404).json({ message: `${id}의 포스트가 없습니다` });
+    }
 }
 
 // 포스트를 작성하는 함수
 export async function createPost(req, res, next) {
-  const { userid, name, text } = req.body;
-  const post = await postRepository.create(userid, name, text);
-  res.status(201).json(post);
+    const { text } = req.body;
+    const post = await postRepository.create(text, req.id);
+    res.status(201).json(post);
 }
 
 // 포스트를 변경하는 함수
 export async function updatePost(req, res, next) {
-  const id = req.params.id;
-  const text = req.body.text;
-  const post = await postRepository.update(id, text);
-  if (post) {
-    res.status(201).json(post);
-  } else {
-    res.status(404).json({ message: `${id}의 포스트가 없습니다` });
-  }
+    const id = req.params.id;
+    const text = req.body.text;
+    const post = await postRepository.getById(id);
+    if (!post) {
+        res.status(404).json({ message: `${id}의 포스트가 없습니다` });
+    }
+    if (post.idx !== req.id) {
+        console.error();
+        return res.sendStatus(403);
+    }
+    const updated = await postRepository.update(id, text);
+    res.status(200).json(updated);
 }
 
 // 포스트를 삭제하는 함수
 export async function deletePost(req, res, next) {
-  const id = req.params.id;
-  await postRepository.remove(id);
-  res.sendStatus(204);
+    const id = req.params.id;
+    const post = await postRepository.getById(id);
+    if (!post) {
+        res.status(404).json({ message: `${id}의 포스트가 없습니다` });
+    }
+    if (post.idx !== req.id) {
+        return res.sendStatus(403);
+    }
+    await postRepository.remove(id);
+    res.sendStatus(204);
 }
